We need to use floats to perform calculations.

+++ assignment3/sequence_generator.py
@@ -10,7 +10,7 @@ def generate_sequence(nb_batch, max_len, batch_size=1, cuda=False):
 
         seq_len = random.randint(2, max_len + 1)
 
-        seq = (numpy.random.rand(seq_len, batch_size, 9) > 0.5).astype(numpy.uint8)
+        seq = (numpy.random.rand(seq_len, batch_size, 9) > 0.5).astype(numpy.float32)
 
         seq[-1] = 0
         seq[:, -1] = 0
